🎨 Move the freecad patch to freecad __init__
